fuse_loop: prevent looper thread stuck on on exit.
fuse_loop implementation loops on events from fuse device, until marked
to exit. Now consider that there is a thead that run the fuse_loop, and
no more events will come from the kernel (no reads, writes or
whatsoever).
This will stall the "join" that will be performed on "looper thread"
after the call to procstat_stop.

So the following scenario can lead to stall:

main:
  1) procstat_create
  2) call procstat_loop on side thread looper1
  3) procstat_stop
  4) join looper1

I could not find any suitable workaround or solution for this.
The only "workaround" worked for me is to "open" root directory before
the call to procstat_stop and "close" it. This of course should yield
the callback into procstat userspace from kernel and successfully exit
the looper thread.